backend/requirements.txt
fastapi
uvicorn
websockets
httpx
python-dotenv
asyncio
livekit-server-sdk
livekit-agents
groq
cartesia
deepgram
silero
pydantic
uuid
requests
json
logging
datetime
time